BlackJack/requirements.txt
streamlit
gym
matplotlib
seaborn
numpy
scikit-learn
tqdm
joblib
gym==0.26.2
numpy==1.23.5
pygame
Pillow